requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
werkzeug==3.0.1
twilio==8.9.0  # For WhatsApp integration
flask-mail==0.9.1  # For email sending
flask-dotenv==0.1.1